Rename advertisements to MixedAdvertisements
