+++ MixedAdvertisements.ts
